requirements.txt
bandit
black
coverage
django-naomi
isort
model-mommy
pre-commit
prospector[with-vulture]
safety
django
celery[redis]
django-model-utils
django-webpack-loader
django-js-reverse
django-import-export
django-ses
python-decouple
psycopg2
brotlipy
django-log-request-id
dj-database-url
gunicorn
whitenoise
psutil
ipython
sentry-sdk
djangorestframework
django-rest-auth
requests
social-auth-core
